# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для загрузки и обработки документов через n8n",
        "title": "Document AI API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "description": "Отдает HTML-страницу с формой загрузки документов",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Home"
                ],
                "summary": "Главная страница",
                "responses": {
                    "200": {
                        "description": "index.html",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Страница не найдена",
                        "schema": {
                            "$ref": "#/definitions/main.APIResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Возвращает статус работы сервиса, время и версию",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Проверка состояния сервиса",
                "responses": {
                    "200": {
                        "description": "Сервис работает нормально",
                        "schema": {
                            "$ref": "#/definitions/main.APIResponse"
                        }
                    }
                }
            }
        },
        "/results": {
            "get": {
                "description": "Возвращает последние результаты обработки документов",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "results"
                ],
                "summary": "Получить список результатов",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.APIResponse"
                        }
                    }
                }
            }
        },
        "/upload": {
            "post": {
                "description": "Загружает файл (PDF, JPG, JPEG, PNG) с описанием и отправляет на обработку через n8n",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "Загрузка документа",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Описание документа",
                        "name": "message",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Файл документа",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Документ успешно отправлен на обработку",
                        "schema": {
                            "$ref": "#/definitions/main.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибки валидации или проблемы с файлом",
                        "schema": {
                            "$ref": "#/definitions/main.APIResponse"
                        }
                    },
                    "405": {
                        "description": "Метод не разрешен",
                        "schema": {
                            "$ref": "#/definitions/main.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера при отправке в n8n",
                        "schema": {
                            "$ref": "#/definitions/main.APIResponse"
                        }
                    }
                }
            }
        },
        "/webhook": {
            "post": {
                "description": "Принимает данные обработки документа от n8n и сохраняет результат",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhook"
                ],
                "summary": "Вебхук от n8n",
                "parameters": [
                    {
                        "description": "Данные от n8n",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Результат успешно сохранен",
                        "schema": {
                            "$ref": "#/definitions/main.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка чтения данных или JSON",
                        "schema": {
                            "$ref": "#/definitions/main.APIResponse"
                        }
                    },
                    "405": {
                        "description": "Метод не разрешен",
                        "schema": {
                            "$ref": "#/definitions/main.APIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}